+++ src/main/java/com/lowdragmc/mbd2/common/gui/editor/MultiblockMachineProject.java
@@ -27,7 +27,6 @@ import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
 import net.minecraft.nbt.Tag;
-import net.minecraft.resources.ResourceLocation;
 
 import java.io.File;
 import java.util.*;