+++ src/main/java/com/lowdragmc/mbd2/integration/kubejs/recipe/MBDRecipeSchema.java
@@ -27,6 +27,7 @@ import com.lowdragmc.mbd2.integration.naturesaura.NaturesAuraRecipeCapability;
 import com.lowdragmc.mbd2.integration.pneumaticcraft.PNCHeatRecipeCapability;
 import com.lowdragmc.mbd2.integration.pneumaticcraft.PNCPressureAirRecipeCapability;
 import com.lowdragmc.mbd2.integration.pneumaticcraft.PressureAir;
+import com.lowdragmc.mbd2.integration.pneumaticcraft.trait.heat.PNCTemperatureCondition;
 import dev.latvian.mods.kubejs.fluid.FluidLike;
 import dev.latvian.mods.kubejs.fluid.FluidStackJS;
 import dev.latvian.mods.kubejs.fluid.InputFluid;
@@ -492,6 +493,14 @@ public interface MBDRecipeSchema {
                 throw new IllegalStateException("Try to add a heat condition while the mekanism is not loaded!");
             }
             addCondition(new MekanismHeatCondition(minHeat, maxHeat));
+            return this;
+        }
+
+        public MBDRecipeJS temperatureCondition(float minTemperature, float maxTemperature) {
+            if (!MBD2.isPneumaticCraftLoaded()) {
+                throw new IllegalStateException("Try to add a temperature condition while the pneumatic is not loaded!");
+            }
+            addCondition(new PNCTemperatureCondition(minTemperature, maxTemperature));
             return this;
         }