+++ CHANGELOG.md
@@ -1,5 +1,3 @@
 # ChangeLog
 
-* fix pnc heat trait won't transfer heat
-* better text display
-* fix existing recipetype project can't set fuel recipe view
+* add temperature condition for PNC recipe